feat(cartridge): introduce Cartridge class
Adds a Cartridge class that is capable of loading a .nes file and selecting the appropriate Mapper from the header. The Cartridge exposes cpu and ppu read/write methods that the CPU and PPU will use to interact with the Cartridge. Some refactoring is done to the test configurations to allow some fixtures to be shared between tests.

1. Adds Cartridge class.
2. Adds a SUPPORTED_MAPPERS constant to the mappers sub-package to expose a list of mappers that are currently supported.
3. Moves the conftest.py used to configure the mapper tests to the tests directory so that fixtures can be shared between tests.
4. Adds test coverage for the Cartridge class.