refactor(cpu): make internal registers and counters public
Refactors the CPU to make the internal registers and counters public. Testing and debugging processes will be required to explicitly set and interrogate these values. Because of the need to set these values, the CPUReadOnlyValues has been removed in favor "Pythonic" getters and setters (making the attributes public). This change also creates two private methods _retrieve_operand and _execute_operation to allow the process of executing addressing modes and operations to be generalized. This is done so that they can easily be patched during testing, removing the need to patch the specific addressing mode or operation mapped to the opcode. Unit tests are updated to account for the changes.