refactor(cpu): distinguish between operand, operation value and effective address
This change renames the publicly exposed operand attribute to operation value. The convention in 6502 literature is to refer to the operand as the immediate value at pc + n used by the operation. The operation value now refers to the value at the address used by the addressing mode. Moreover, a private _effective_address value is added to store the address of the operation_value. This will be required by operations that manipulate the value stored at the effective address. Tests are updated to account for this change.